+++ tests/test_sheet_update.py
@@ -25,20 +25,6 @@ class TestReader:
         r.set_named_column_at('b', [11, 1])
         assert r.column_at(1) == [11, 1]
 
-    def test_update_a_cell_with_a_filter(self):
-        """
-        Filter the sheet first and then update the filtered now
-
-        with the filter, you can set its value. then clear
-        the filters, the value stays with the cell. so if you want
-        to save the change with original data, please clear the filter
-        first
-        """
-        r = pe.get_sheet(file_name=self.testfile)
-        r.add_filter(pe.sheets.filters.ColumnFilter([0, 2]))
-        r.cell_value(2, 1, "k")
-        assert r[2, 1] == "k"
-
     def tearDown(self):
         clean_up_files([self.testfile])
 
@@ -103,91 +89,6 @@ class TestSeriesReader:
         assert r2.row[1] == ['e', 'f', 'g', 'h', 'x1', 'x2', 'x4']
         assert r2.row[2] == ['i', 'j', 1.1, 1, 'y1', 'y2', '']
         assert r2.row[3] == ['', '', '', '', 'z1', '', '']
-
-    def tearDown(self):
-        clean_up_files([self.testfile])
-
-
-class TestReaderWithFilter:
-    def setUp(self):
-        """
-        Make a test csv file as:
-
-        1, 2, 3, 4
-        5, 6, 7, 8
-        9, 10,11,12
-        """
-        self.testfile = "test.xlsm"
-        create_sample_file2(self.testfile)
-
-    def test_add_rows_even_row_filter(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        r.add_filter(pe.sheets.filters.EvenRowFilter())
-        assert r.number_of_rows() == 2
-        assert r.number_of_columns() == 4
-        result = [1, 2, 3, 4, 9, 10, 11, 12]
-        actual = list(r.enumerate())
-        eq_(result, actual)
-
-    def test_add_rows_even_row_filter_2(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        content = [['r', 's', 't', 'o'],  # 4
-                   [1, 2, 3, 4],  # 5
-                   [True],  # 6
-                   [1.1, 2.2, 3.3, 4.4, 5.5]]  # 7
-        r.extend_rows(content)
-        r.add_filter(pe.sheets.filters.EvenRowFilter())
-        eq_(r.row[3], content[3])
-
-    def test_delete_rows_even_row_filter(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        r.add_filter(pe.sheets.filters.EvenRowFilter())
-        assert r.number_of_rows() == 2
-        assert r.number_of_columns() == 4
-        result = [1, 2, 3, 4, 9, 10, 11, 12]
-        actual = list(r.enumerate())
-        eq_(result, actual)
-        del r.row[0]
-        result = [5, 6, 7, 8]
-        actual = list(r.enumerate())
-
-    def test_add_rows_odd_column_filter(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        r.add_filter(pe.sheets.filters.OddColumnFilter())
-        assert r.number_of_rows() == 3
-        assert r.number_of_columns() == 2
-        result = [2, 4, 6, 8, 10, 12]
-        actual = list(r.enumerate())
-        eq_(result, actual)
-
-    def test_add_rows_odd_column_filter_2(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        #             5     6    7
-        rows = [['c1', 'c2', 'c3'],
-                ['x1', 'x2', 'x4']]
-        r.extend_columns(pe.sheets.transpose(rows))
-        r.add_filter(pe.sheets.filters.OddColumnFilter())
-        assert r.row[0] == [2, 4, 'c2']
-        assert r.row[1] == [6, 8, 'x2']
-        assert r.row[2] == [10, 12, '']
-
-    def test_delete_rows_odd_column_filter(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        r.add_filter(pe.sheets.filters.OddColumnFilter())
-        assert r.number_of_rows() == 3
-        assert r.number_of_columns() == 2
-        result = [2, 4, 6, 8, 10, 12]
-        actual = list(r.enumerate())
-        eq_(result, actual)
-
-    def test_delete_rows_odd_column_filter_2(self):
-        r = pe.get_sheet(file_name=self.testfile)
-        #             5     6    7
-        r.delete_columns([0])
-        r.add_filter(pe.sheets.filters.OddColumnFilter())
-        result = [3, 7, 11]
-        actual = list(r.enumerate())
-        eq_(result, actual)
 
     def tearDown(self):
         clean_up_files([self.testfile])